+++ src/90-subsetsWithDup.cpp
@@ -19,7 +19,7 @@ public:
 private:
     void dfs(vector<int>& nums, int index, vector<int>&& path, vector<vector<int>>& res) {
         res.emplace_back(path);
-        for (int i = index; i < nums.size(); ++i) {
+        for (int i = index; i < static_cast<int>(nums.size()); ++i) {
             if (i > index && nums[i] == nums[i - 1]) {
                 continue;
             }
